[committed][RISC-V][PR target/114139] Verify we have a CONST_INT before extracting INTVAL
Run-of-the-mill checking issue.  We had something like (plus (reg) (reg)) and
tried to extract INTVAL (XEXP (x, 1)) which of course blows up with checking
on.

Fixed thusly.   Tested on riscv32-elf in my tester.  riscv64-elf is in flight,
but won't finish for a while due to other tasks in flight.

	PR target/114139
gcc/
	* config/riscv/riscv.cc (riscv_macro_fusion_pair_p): Verify object
	is a CONST_INT before looking at INTVAL.

gcc/testsuite/

	* gcc.target/riscv/pr114139.c: New test.