contracts: Allow disabling optimization of checking fns
P2680 shows a situation where optimization of contracts expressions that
exploit assumption that UB cannot occur - by completely eliding checks in
those cases; which actually prevents the contract checking to work as the
user intended.

As a proof-of-principle, this patch allows use to disable optimization of
the contract checks.  It is not proposed as a production solution - really
we need to figure out the proper course of action.