[committed] Fix fr30-elf newlib build failure with late-combine
So the late combine work has exposed a latent bug in the fr30 port.

The fr30 "call" instruction is pc-relative with a *very* limited range, 12 bits
to be precise.

With such a limited range its hard to see how we could ever consistently use it
in the compiler, with the possible exception of self-recursion.  Even for a
call to a locally binding function -ffunction-sections and linker placement of
functions may separate the caller/callee.  Code generation seemed to be using
indirect forms pretty consistently, though the RTL would allow direct calls.

With late-combine some of those indirects would be optimized into direct calls.
This naturally led to out of range scenarios.

With the fr30 port slated for removal unless it gets updated to use LRA and the
fundamental problems using direct calls, I took the shortest path to keep
things working -- namely forcing all calls to be indirect.

Tested in my tester with no regressions (and fixes the newlib build failure
with late-combine enabled).    Pushed to the trunk.

gcc/
	* config/fr30/constraints.md (Q): Remove unused constraint.
	* config/fr30/predicates.md (call_operand): Remove unused predicate.
	* config/fr30/fr30.md (call, vall_value): Turn into expanders and
	force the call address into a register.
	(*call, *call_value): Adjust to only allow indirect calls.  Adjust
	output template accordingly.