[committed][RISC-V] Fix expected output for thead store pair test
Surya's patch to IRA has improved the code we generate for one of the thead
store pair tests for both rv32 and rv64.  This patch adjusts the expectations
of that test.

I've verified that the test now passes on rv32 and rv64 in my tester.  Pushing
to the trunk.

gcc/testsuite
	* gcc.target/riscv/xtheadmempair-3.c: Update expected output.