ada: Fix internal error on limited aggregate in nested conditional expression
This is a fallout of an earlier fix to Is_Build_In_Place_Aggregate_Return
that made it take into account intermediate conditional expressions, but
turned out to work only for a single nesting level of them.

The fix reuses the delayed expansion mechanism recently extended to deal
with conditional expressions in a straightforward way.

gcc/ada/

	* exp_aggr.adb (Convert_To_Assignments): Set Expansion_Delayed on
	intermediate conditional expressions for BIP aggregate returns
	too.
	* exp_ch4.adb (Expand_N_Case_Expression): Also deal with delayed
	expansion in the return case.
	(Expand_N_If_Expression): Likewise.