jit: Fix Darwin bootstrap after r15-1699.
r15-1699-g445c62ee492 contains changes that trigger two maybe-uninitialized
warnings on Darwin, which result in a bootstrap failure.

Note that the warnings are false positives, in fact the variables should be
initialized in the cases of a switch (all values of the switch condition are
covered).

Fixed here by providing default initializations for the relevant variables.

gcc/jit/ChangeLog:

	* jit-recording.cc
	(recording::memento_of_typeinfo::make_debug_string): Default the value
	of ident.
	(recording::memento_of_typeinfo::write_reproducer): Default the value
	of type.

Signed-off-by: Iain Sandoe <iains@gcc.gnu.org>