Preserve SSA info for more propagated copy
Besides VN and copy-prop also CCP and VRP as well as forwprop
propagate out copies and thus it's worthwhile to try to preserve
range and points-to info there when possible.

Note that this also fixes the testcase from PR115701 but that's
because we do not actually intersect info but only copy info when
there was no info present.

	* tree-ssa-forwprop.cc (fwprop_set_lattice_val): Preserve
	SSA info.
	* tree-ssa-propagate.cc
	(substitute_and_fold_dom_walker::before_dom_children): Likewise.