sparc: define SPARC_LONG_DOUBLE_TYPE_SIZE for vxworks [PR115739]
Commit r15-1594 removed define of LONG_DOUBLE_TYPE_SIZE in
sparc.cc, it's based on the assumption that each OS has its
own define (see the comments in sparc.h), but it exposes an
issue on vxworks which lacks of the define.

We can bring back the default SPARC_LONG_DOUBLE_TYPE_SIZE to
sparc.cc, but according to the comments in sparc.h, I think
it's better to define this in vxworks.h.  btw, I also went
through all the sparc supported triples, vxworks is the only
one that misses this define.

	PR target/115739

gcc/ChangeLog:

	* config/sparc/vxworks.h (SPARC_LONG_DOUBLE_TYPE_SIZE): New define.