c++: array new with value-initialization [PR115645]
This extends the r11-5179 fix which doesn't work with multidimensional
arrays.  In particular,

  struct S {
    explicit S() { }
  };
  auto p = new S[1][1]();

should not say "converting to S from initializer list would use
explicit constructor" because there's no {}.  However, since we
went into the block where we create a {}, we got confused.  We
should not have gotten there but we did because array_p was true.

This patch refines the check once more.

	PR c++/115645

gcc/cp/ChangeLog:

	* init.cc (build_new): Don't do any deduction for arrays with
	bounds if it's value-initialized.

gcc/testsuite/ChangeLog:

	* g++.dg/expr/anew7.C: New test.